% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_mail.R
\name{ops_mail}
\alias{ops_mail}
\title{Send an operational email}
\usage{
ops_mail(mgs, subject, from, to = NULL)
}
\arguments{
\item{from}{The sender. If NULL, defaults to the recipient specified in "to"}

\item{to}{The recipient.}

\item{msg}{A string containing the markdown message you wish to send}
}
\description{
Used to warn of process success or failure. Use sparingly.
}
