% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_slack.R
\name{ops_slack}
\alias{ops_slack}
\title{Send messages to slack}
\usage{
ops_slack(msg, webhook_url, complex_msg = NULL)
}
\arguments{
\item{msg}{A basic character string with your message}

\item{webhook_url}{Your Slack App webhook's incoming URL}

\item{complex_message}{A more complex message as a list (will be converted
to JSON)}
}
\description{
Using an incoming webhook, sends a message to slack
}
