% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryCatch.R
\name{ops_tc}
\alias{ops_tc}
\title{tryCatch Wrapper}
\usage{
ops_tc(expr, on_error, ...)
}
\arguments{
\item{expr}{An R expression to try}

\item{on_error}{An R expression to run if there was an error}
}
\description{
A simple tryCatch wrapper that checks if a result is an error and then
performs some action. Useful for appending a list of failed ids to a
log file.
}
