% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{soma}
\alias{soma}
\title{Addition}
\usage{
soma(a, b)
}
\arguments{
\item{a}{number}

\item{b}{number}
}
\value{
number
}
\description{
In addition there is the calculus of adding natural numbers to others.
This mathematical operation is also popularly known as addition. The
final result of the addition is called the total or sum and the numbers used are
the parcels. The arithmetic operator, that is, the sign that indicates
its calculation is the (+)
}
\examples{
soma(9,2)
}
